# api/openapi.yaml
openapi: 3.1.0
info:
  title: Hookyard API
  version: 0.1.0
  summary: Reliable delivery for outbound API calls.
  description: |
    Hookyard delivers outbound HTTP requests to third-party APIs ("upstreams") on behalf of your
    applications, with retries, backoff and a dead-letter queue.

    ## Concepts

    - **Upstream**: a third-party API configured in Hookyard (base URL, credentials, retry policy).
      Applications refer to upstreams by name and never handle vendor URLs or credentials.
    - **Request**: one outbound call that an application asked Hookyard to deliver.
    - **Attempt**: one HTTP try of a request.
    - **Dead-letter queue (DLQ)**: requests in the `dead` state that exhausted their retries or failed
      permanently. They can be inspected and replayed.

    ## Conventions

    - **Authentication**: every `/v1` endpoint requires `Authorization: Bearer <token>`.
    - **Durations** are strings with a unit, for example `"500ms"`, `"30s"`, `"5m"` or `"2h"`.
    - **Timestamps** are RFC 3339 in UTC.
    - **Errors** always use the `Error` schema, and `error.code` is stable and safe to branch on.
    - **Pagination** is cursor based: pass `next_cursor` from a response as `cursor` to get the next page.

    ## Delivery

    - Network errors, timeouts, `408`, `425`, `429` and `5xx` responses are retried with exponential
      backoff and full jitter. A `Retry-After` header on `429` or `503` sets the delay.
...
      moves to the dead-letter queue immediately.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /v1/requests:
    post:
      operationId: createRequest
      tags: [Requests]
      summary: Enqueue a request
      description: |
        Queues an outbound request for delivery and returns immediately.

        If `dedupe_key` is set and a request with the same key already exists for the same upstream
        (within the dedupe window, 24 hours by default), no new request is created. The existing request
        is returned with status `200` and the `Hookyard-Deduplicated: true` header. This makes it
        safe for applications to retry the enqueue call itself.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequest"
            examples:
              minimal:
                summary: Minimal request
                value:
                  upstream: courier-x
                  method: POST
                  path: /shipments
                  body: { order_id: 123 }
              full:
                summary: All options
                value:
                  upstream: courier-x
                  method: POST
                  path: /shipments?notify=true
                  headers: { X-Correlation-Id: 7f3c1e }
                  body: { order_id: 123, weight_kg: 2.5 }
                  dedupe_key: order-123-create-shipment
                  deliver_at: "2026-10-01T10:00:00Z"
                  timeout: 15s
                  retry: patient
                  tags: { app: orders, tenant: acme }
      responses:
        "202":
          description: The request was queued.
          headers:
            Location:
              description: URL of the created request.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "200":
          description: A request with the same `dedupe_key` already exists; it is returned unchanged.
          headers:
            Hookyard-Deduplicated:
              description: Always `true` for this response.
              schema: { type: string, enum: ["true"] }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listRequests
      tags: [Requests]
      summary: List requests
      description: Lists requests, newest first. All filters are combined with AND.
      parameters:
        - $ref: "#/components/parameters/Upstream"
        - name: status
          in: query
          description: Filter by one or more statuses, comma separated.
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/RequestStatus"
          example: [dead, failed]
        - name: tag
          in: query
          description: Filter by tag, as `key:value`. Repeat the parameter to match several tags.
          style: form
          explode: true
          schema:
            type: array
            items: { type: string, pattern: "^[^:]+:.*$" }
          example: ["app:orders"]
        - name: dedupe_key
          in: query
          description: Filter by exact dedupe key.
          schema: { type: string }
        - $ref: "#/components/parameters/CreatedAfter"
        - $ref: "#/components/parameters/CreatedBefore"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of requests.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/requests/{id}:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    get:
      operationId: getRequest
      tags: [Requests]
      summary: Get a request
      responses:
        "200":
          description: The request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/requests/{id}/attempts:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    get:
      operationId: listAttempts
      tags: [Requests]
      summary: List delivery attempts
      description: Returns every delivery attempt of a request, oldest first.
      responses:
        "200":
          description: The attempts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/requests/{id}/replay:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    post:
      operationId: replayRequest
      tags: [Requests]
      summary: Replay a request
      description: |
        Queues a finished request (`dead`, `succeeded`, `canceled` or `unknown`) for delivery again with a
        fresh retry budget. Previous attempts are kept. Replaying a request that is still in progress
        returns `409`.
      responses:
        "202":
          description: The request was queued again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/requests/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    post:
      operationId: cancelRequest
      tags: [Requests]
      summary: Cancel a request
      description: |
        Cancels a request that has not finished (`scheduled`, `pending` or `failed`). A request that is
        `in_flight` cannot be canceled, because the call to the upstream may already have been sent.
      responses:
        "200":
          description: The request was canceled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/requests/{id}/resolve:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    post:
      operationId: resolveRequest
      tags: [Requests]
      summary: Settle a request whose outcome is unknown
      description: |
        A request is `unknown` when it was sent but no response arrived, so Hookyard can't tell whether
        the upstream processed it. After checking with the upstream, record what happened: `succeeded`,
        or `dead` (it didn't go through; it moves to the dead-letter queue, where it can be replayed).
        Only `unknown` requests can be resolved. Recorded in the audit log.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [outcome]
              properties:
                outcome:
                  type: string
                  enum: [succeeded, dead]
                reason:
                  type: string
                  description: Why, for the audit log.
            example:
              outcome: succeeded
              reason: The vendor's dashboard shows the charge.
      responses:
        "200":
          description: The resolved request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/requests/{id}/callbacks:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    get:
      operationId: listCallbacks
      tags: [Callbacks]
      summary: List completion callbacks
      description: |
        Returns the callbacks sent (or being sent) for a request, oldest first. A request gets one each
        time it reaches a final state with a callback URL set; a replayed request can have several.
      responses:
        "200":
          description: The callbacks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallbackList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/requests/{id}/callbacks/{callback_id}/retry:
    parameters:
      - $ref: "#/components/parameters/RequestId"
      - name: callback_id
        in: path
        required: true
        schema:
          type: string
          example: evt_0f6a3c2b9d8e4f5a8b7c6d5e4f3a2b1c
    post:
      operationId: retryCallback
      tags: [Callbacks]
      summary: Retry a failed callback
      description: |
        Sends a `failed` callback again, with a fresh attempt budget. The event (its id and body) is
        unchanged, so receivers that deduplicate on `webhook-id` handle it once. Recorded in the audit
        log.
      responses:
        "202":
          description: The callback was queued again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Callback"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/dlq:
    get:
      operationId: getDLQSummary
      tags: [Dead-letter queue]
      summary: Summarize the dead-letter queue
      description: |
        Groups dead requests by upstream and failure reason. Use `GET /v1/requests?status=dead` to list
        the individual requests.
      parameters:
        - $ref: "#/components/parameters/Upstream"
      responses:
        "200":
          description: DLQ groups, largest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DLQSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/dlq/replay:
    post:
      operationId: replayDLQ
      tags: [Dead-letter queue]
      summary: Replay dead requests in bulk
      description: |
        Queues every dead request that matches the filter for delivery again. Filters are combined with
        AND. An empty filter matches the whole DLQ, so use `dry_run` first to see how many requests
        would be replayed. The operation is idempotent: requests that were already replayed are no
        longer dead and are not matched again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DLQReplayRequest"
            examples:
              byReason:
                summary: Replay timeouts for one upstream from the last two hours
                value:
                  upstream: courier-x
                  error_code: timeout
                  dead_after: "2026-09-29T08:00:00Z"
              dryRun:
                summary: Count without replaying
                value:
                  upstream: courier-x
                  dry_run: true
      responses:
        "200":
          description: The replay result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DLQReplayResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/upstreams:
    get:
      operationId: listUpstreams
      tags: [Upstreams]
      summary: List upstreams
      responses:
        "200":
          description: All configured upstreams, sorted by name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/upstreams/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: Upstream name.
        schema:
          $ref: "#/components/schemas/UpstreamName"
    get:
      operationId: getUpstream
      tags: [Upstreams]
      summary: Get an upstream
      description: Returns the effective configuration of an upstream. Secret values are never returned.
      responses:
        "200":
          description: The upstream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upstream"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/upstreams/{name}/pause:
    parameters:
      - $ref: "#/components/parameters/UpstreamName"
    post:
      operationId: pauseUpstream
      tags: [Upstreams]
      summary: Pause deliveries to an upstream
      description: |
        Requests to a paused upstream wait in the queue: they don't use up attempts, and the paused time
        doesn't count against their `max_age`. The pause survives restarts. Set `until` or `duration`
        to end it automatically; pausing again updates the reason and end time. Recorded in the audit
        log and the upstream's events.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason:
                  type: string
                  maxLength: 500
                until:
                  type: string
                  format: date-time
                  description: End the pause at this time (at most 30 days ahead).
                duration:
                  $ref: "#/components/schemas/Duration"
            example:
              reason: Vendor maintenance window
              duration: 2h
      responses:
        "200":
          description: The upstream, now paused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upstream"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/upstreams/{name}/resume:
    parameters:
      - $ref: "#/components/parameters/UpstreamName"
    post:
      operationId: resumeUpstream
      tags: [Upstreams]
      summary: Resume deliveries to a paused upstream
      description: Waiting requests are delivered again, with the paused time credited to their `max_age`.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason: { type: string }
      responses:
        "200":
          description: The upstream, resumed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upstream"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/upstreams/{name}/events:
    parameters:
      - $ref: "#/components/parameters/UpstreamName"
    get:
      operationId: listUpstreamEvents
      tags: [Upstreams]
      summary: Upstream history
      description: Circuit breaker transitions, pauses and resumes, newest first.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        "200":
          description: Events, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UpstreamEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/stats/overview:
    get:
      operationId: getStatsOverview
      tags: [Stats]
      summary: Per-upstream health overview
      parameters:
        - name: window
          in: query
          description: How far back to aggregate.
          schema:
            $ref: "#/components/schemas/Duration"
            default: 1h
      responses:
        "200":
          description: One entry per upstream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/stats/timeseries:
    get:
      operationId: getStatsTimeseries
      tags: [Stats]
      summary: Delivery metrics over time
      parameters:
        - $ref: "#/components/parameters/Upstream"
        - name: from
          in: query
          description: Start of the range. Defaults to one hour before `to`.
          schema: { type: string, format: date-time }
        - name: to
          in: query
          description: End of the range. Defaults to now.
          schema: { type: string, format: date-time }
        - name: step
          in: query
          description: |
            Bucket size, rounded up to whole minutes. Buckets start at `from` (rounded down to the
            minute). A range may have at most 1500 buckets.
          schema:
            $ref: "#/components/schemas/Duration"
            default: 1m
      responses:
        "200":
          description: Time buckets, oldest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsTimeseries"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API token configured in Hookyard.
...
  responses:
    BadRequest:
      description: |
        The request is malformed: invalid JSON, unknown fields or bad query parameters. Request
        bodies are always parsed as JSON, whatever the `Content-Type` header says.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: bad_request
              message: "invalid query parameter \"limit\": must be between 1 and 200"
    Unauthorized:
      description: The bearer token is missing or invalid.
      headers:
        WWW-Authenticate:
          description: Always `Bearer realm="hookyard"`.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: unauthorized, message: missing or invalid bearer token }
...
    PayloadTooLarge:
      description: The request body exceeds the configured limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: payload_too_large, message: request body exceeds 1 MiB }
    UnprocessableEntity:
      description: The request is well formed but fails validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            unknownUpstream:
              summary: Unknown upstream
              value:
                error:
                  code: unknown_upstream
                  message: "upstream \"courier-y\" is not configured (did you mean \"courier-x\"?)"
            validation:
              summary: Invalid fields
              value:
                error:
                  code: validation_failed
                  message: 2 fields are invalid
                  details:
                    - { field: path, message: "must start with \"/\"" }
                    - { field: retry.max_attempts, message: must be between 1 and 100 }
    InternalError:
      description: An unexpected server error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: internal, message: internal server error }
  schemas:
...
    HTTPMethod:
      type: string
      enum: [GET, POST, PUT, PATCH, DELETE]
...
    CreateRequest:
      type: object
      additionalProperties: false
      required: [upstream, method, path]
      properties:
        upstream:
          $ref: "#/components/schemas/UpstreamName"
        method:
          $ref: "#/components/schemas/HTTPMethod"
        path:
          type: string
          description: Path (and optional query string) appended to the upstream's base URL.
          pattern: "^/"
          maxLength: 2048
          example: /shipments?notify=true
        headers:
          allOf:
            - $ref: "#/components/schemas/Headers"
          description: |
            Headers for this request. Headers configured on the upstream (typically credentials) take
            precedence. `Host`, `Content-Length` and hop-by-hop headers can't be set.
        body:
          description: |
            Request body. A JSON object, array, number or boolean is sent as JSON with
            `Content-Type: application/json`. A string is sent verbatim, so set a `Content-Type` header
            for other formats (for example XML or form data). Omit it, or use `null`, to send no body.
        dedupe_key:
          type: string
          description: |
            Makes enqueueing idempotent. While a request with the same key exists for the same upstream
            (within the dedupe window), enqueueing again returns that request instead of creating a new
            one.
          minLength: 1
          maxLength: 256
          example: order-123-create-shipment
        deliver_at:
          type: string
          format: date-time
          description: Deliver no earlier than this time. The request stays `scheduled` until then.
        timeout:
          allOf:
            - $ref: "#/components/schemas/Duration"
          description: Per-attempt timeout. Overrides the upstream's timeout.
        retry:
          $ref: "#/components/schemas/RetryPolicy"
        tags:
          $ref: "#/components/schemas/Tags"
        callback_url:
          type: string
          format: uri
          maxLength: 2048
          description: |
            Where to POST a signed event when the request finishes (see the `requestFinished` webhook).
            Overrides the upstream's `callback_url`; an empty string turns callbacks off for this request.
            Must match `callbacks.allow` in `hookyard.yaml` when that is set, and requires
            `HOOKYARD_CALLBACK_SECRETS` on the server.
          example: http://orders-svc:3000/hooks/hookyard
        on_result:
          type: string
          minLength: 1
          maxLength: 128
          description: An app-defined key carried in the callback event, to route it to a handler.
          example: order.shipment
    Request:
      type: object
      required: [id, upstream, method, path, status, attempt_count, retry, tags, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/RequestId"
        upstream:
          $ref: "#/components/schemas/UpstreamName"
        method:
          $ref: "#/components/schemas/HTTPMethod"
        path:
          type: string
          example: /shipments
        headers:
          $ref: "#/components/schemas/Headers"
        body:
          description: The body as it was enqueued.
        dedupe_key:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/RequestStatus"
        attempt_count:
          type: integer
          description: Number of attempts made so far.
          minimum: 0
        retry:
          $ref: "#/components/schemas/EffectiveRetryPolicy"
        timeout:
          $ref: "#/components/schemas/Duration"
        tags:
          $ref: "#/components/schemas/Tags"
        deliver_at:
          type: [string, "null"]
          format: date-time
        next_attempt_at:
          type: [string, "null"]
          format: date-time
          description: When the next attempt is due, for `scheduled`, `pending` and `failed` requests.
        last_error:
          oneOf:
            - $ref: "#/components/schemas/DeliveryError"
            - type: "null"
          description: Why the most recent attempt failed, if it did.
        last_status_code:
          type: [integer, "null"]
          description: HTTP status code of the most recent response, if any.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
          description: When the request reached a final state.
        callback_url:
          type: [string, "null"]
          description: Where the completion callback goes, or `null` for none.
        on_result:
          type: [string, "null"]
          description: The routing key carried in the callback event.
...